seperate CinnLaunchContext class from cinn_launch_op.h and put it in a new independent file
