+++ paddle/fluid/operators/cinn/cinn_launch_op.h
@@ -26,6 +26,7 @@
 #include "paddle/fluid/framework/op_registry.h"
 #include "paddle/fluid/framework/operator.h"
 #include "paddle/fluid/framework/paddle2cinn/cinn_compiler.h"
+#include "paddle/fluid/operators/cinn/cinn_launch_context.h"
 
 namespace paddle {
 namespace operators {
@@ -41,69 +42,6 @@ using CinnCompiler = framework::paddle2cinn::CinnCompiler;
 using CinnCompiledObject = framework::paddle2cinn::CinnCompiledObject;
 
 namespace details {
-
-class CinnLaunchContext {
- public:
-  explicit CinnLaunchContext(const CinnCompiledObject& compiled_obj);
-
-  // Return whether a Paddle variable used on compiled kernels
-  bool IsVariableUsed(const std::string& var_name);
-
-  // Assign tensor buffer to input or output variables
-  void AssignExternalVariable(const std::string& var_name,
-                              const platform::Place& place, LoDTensor* tensor);
-
-  // Assign tensor buffer to internal variables
-  void AssignInternalVariable(const std::string& var_name,
-                              const platform::Place& place, LoDTensor* tensor);
-
-  // Extract internal variable names from CinnScope
-  // by excluding used input and output variables
-  std::unordered_set<std::string> GetInternalVariableNames();
-
-  // Finalize all execution arguments and return them
-  const std::map<std::string, cinn_pod_value_t>& FinalizeArguments() const;
-
-  std::vector<std::unique_ptr<cinn_buffer_t>> HandoverBuffers() {
-    return std::move(hold_buffers_);
-  }
-
- private:
-  // Get CinnTensor with CINN variable name
-  CinnTensor GetCinnTensor(const std::string& var_name);
-
-  // Check whether tensors from Paddle and CINN of the same variable
-  // are equivalent in type and dimension
-  void CheckTensorEquivalent(const std::string& var_name,
-                             const LoDTensor& paddle_tensor,
-                             const CinnTensor& cinn_tensor);
-
-  // Share the buffer of a Paddle tensor to CINN by delivering memory address
-  // to a cinn_buffer_t object
-  std::unique_ptr<cinn_buffer_t> ShareTensorWithCinnBuffer(
-      const platform::Place& place, bool free_mem_callback, LoDTensor* tensor);
-
-  // Set an argument with (cinn name)->(paddle tensor) pair
-  void SetArgument(const std::string& cinn_name, const platform::Place& place,
-                   bool free_mem_callback, LoDTensor* paddle_tensor);
-
- private:
-  // a variable name map from paddle to cinn
-  const std::unordered_map<std::string, std::string>& paddle2cinn_varmap_;
-  // the variable scope of cinn
-  const std::shared_ptr<CinnScope> cinn_scope_;
-
-  // all variables used by compiled executable program
-  std::unordered_set<std::string> cinn_variable_names_;
-
-  // because a cinn_pod_value_t does not own the cinn_buffer_t object,
-  // an extra stroage is necessary to keep the object and it can
-  // not be released until runtime program finish  execution.
-  std::vector<std::unique_ptr<cinn_buffer_t>> hold_buffers_;
-
-  // name to execution argument
-  std::map<std::string, cinn_pod_value_t> name2argument_;
-};
 
 // Tranform Paddle place to CINN target
 const ::cinn::common::Target& PlaceToCinnTarget(const platform::Place& place);
@@ -178,8 +116,8 @@ class CinnLaunchOpKernel : public framework::OpKernel<T> {
         compilation_key, inputs_name2tensor, target, stream);
     details::DebugCinnCompiledResult(cinn_compiled_object);
 
-    auto launch_context =
-        std::make_unique<details::CinnLaunchContext>(cinn_compiled_object);
+    auto launch_context = std::make_unique<details::CinnLaunchContext>(
+        cinn_compiled_object.paddle2cinn_varmap, cinn_compiled_object.scope);
 
     // Step 3. Prepare arguments needed for the compiled executable program.
     VLOG(4) << "CinnLaunchOp prepare arguments";
